external-api: types: Move supported tokens from behind feature flag
