+++ external-api/src/types/mod.rs
@@ -20,3 +20,24 @@ pub use api_order_book::*;
 pub use api_task_history::*;
 #[cfg(feature = "wallet-api")]
 pub use api_wallet::*;
+use serde::{Deserialize, Serialize};
+
+// -------------
+// | API Types |
+// -------------
+
+/// A token in the the supported token list
+#[derive(Clone, Debug, Serialize, Deserialize)]
+pub struct ApiToken {
+    /// The token address
+    pub address: String,
+    /// The token symbol
+    pub symbol: String,
+}
+
+impl ApiToken {
+    /// Constructor
+    pub fn new(addr: String, sym: String) -> Self {
+        Self { address: addr, symbol: sym }
+    }
+}
